+++ backend/fms_core/viewsets/process_measurement.py
@@ -1,4 +1,4 @@
-from django.db.models import Q, F, Count, Case, When, IntegerField
+from django.db.models import Q, F, Count
 from django.forms import ValidationError
 
 from rest_framework import viewsets
@@ -100,18 +100,3 @@ class ProcessMeasurementViewSet(viewsets.ModelViewSet, TemplateActionsMixin):
                 for c in self.queryset.values("process__protocol").annotate(Count("process__protocol"))
             },
         })
-    
-    @action(detail=False, methods=["get"])
-    def last_protocols(self, _request):
-        samples = [int(x) for x in self.request.query_params.get("samples").split(",")]
-        queryset = (self.get_queryset()
-                    .annotate(protocol=F('process__protocol__name'))
-                    .annotate(sample_result=Case(When(Q(source_sample__in=samples) & Q(child_sample=None), then=F('source_sample')), # QC?
-                                                 When(Q(child_sample__in=samples), then=F('child_sample')), # child?
-                                                 default=None,
-                                                 output_field=IntegerField()))
-                    .exclude(sample_result=None)
-                    .order_by('sample_result', '-pk')
-                    .distinct('sample_result')
-                    .values('sample_result', 'protocol'))
-        return Response(queryset)
